storefront coloring
I fixed the coloring of the business storefronts! If you open the html file now, you see an example of colors I set myself. To change the storefront colors, all you have to do is go to the lines that say <!--Change color here!--> and change the hex color code.

I think this could be used in Hannah's carousel.